backend/requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.21
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
Flask-CORS==4.0.0
Pillow==10.0.1
torch==2.0.1
transformers==4.33.2
requests==2.31.0
